Add pre set event
